# server/app/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Catalyst-core BOTS",
        "description": "Documentation for catalyst-core BOT API's",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3001/"
        },
        {
            "url": "https://localhost:3001/"
        }
    ],
...
    "paths": {
        "/auth/signin": {
            "post": {
                "tags": [
                    "AUTHENTICATION"
                ],
                "summary": "Authentication of user",
                "requestBody": {
                    "description": "Authentication details to login",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/auth"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Authentication successfull",
                        "content": {}
                    },
                    "500": {
                        "description": "Authentication failed",
                        "content": {}
                    }
                },
                "x-codegen-request-body-name": "body"
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "auth": {
                "required": [
                    "RequestBody"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "relevancelab"
                    },
                    "pass": {
                        "type": "string",
                        "example": "relevancelab@123"
                    },
                    "authType": {
                        "type": "string",
                        "example": "token"
                    }
                }
            }
        }
    }
}